chore: refactor vfkit
Move vfkit launching to its own file. No changes to implementation.
